### 2017-01-30 #96 @adamhsparks
getCRUCLdata: Download and Use CRU CL2.0 Climatology Data in R
The getCRUCLdata package provides two functions that automate downloading and importing CRU CL2.0 climatology data, facilitates the calculation of minimum temperature and maximum temperature, and formats the data into a tidy data frame or a list of raster stack objects for use in an R session or easily exports to a raster format file for use in a geographic information system (GIS).

```
Package: getCRUCLdata
Title: Download and Use CRU CL2.0 Climatology Data in R
Version: 0.1.2
Authors@R: person(given = "Adam", family = "Sparks", email = "adamhsparks@gmail.com", role = c("aut", "cre"))
Description: Provides functions that automate downloading and importing
    University of East Anglia Climate Research Unit (CRU) CL2.0 climatology data
    into R, facilitates the calculation of minimum temperature and maximum
    temperature and formats the data into a tidy data frame or a list of raster
    stack objects for use in an R session.  CRU CL2.0 data are a gridded
    climatology of 1961-1990 monthly means released in 2002 and cover all land
    areas (excluding Antarctica) at 10-minute resolution.  For more
    information see the description of the data provided by the University of
    East Anglia Climate Research Unit,
    <https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt>.
    curl,
    plyr,
    purrr (>= 0.2.0),
    raster,
    utils
    covr
```

https://github.com/adamhsparks/getCRUCLdata

Anyone interested in using climate data or generating raster files from the CRU CL2.0 data that are easy to use in a GIS

Not that I'm aware of

    - [x] The package is deposited in a long-term repository with the  DOI: https://doi.org/10.5281/zenodo.264405

Ivan Hanigan, https://github.com/ivanhanigan
