### 2017-01-30 95447f6
Release 0.1.3, add to Zenodo for DOI
