### 2017-01-30 ffbcdae
Revert "Release 0.1.3, add to Zenodo for DOI"
This reverts commit 95447f6d08e8e6bd554e3482946895a4f02f6527.
