### 2017-01-31 c12dcea
check that tmx calculation works properly
