### 2017-01-31 86f6a69
Check that elv is properly reformatted
