### 2017-01-31 1f0911f
Add check for elevation stack as well
