### 2017-01-31 #96 @sckott
### Editor checks:

---

#### Editor comments

Currently seeking reviewers. It's a good fit and not overlapping.

* Tests do succeed, but take a long time, would be nice to get these to run faster somehow
* Ran `goodpractice::gp()` , but nothing to deal with at this time.

---

Reviewers: xxx
Due date: xxx
