### 2017-02-01 #96 @adamhsparks
@sckott, I'm working on getting the test time down. I've managed to reduce it by 1/2 already and keep the code coverage the same. I'll see what I can do further.
