### 2017-02-01 b5a98d1
Bug fix and new test
Raster layers were incorrectly cropped, extent xmin/xmax and ymin/ymax
numbers were transposed this fixes that bug

Also update test for elv raster layer to be faster
