+++ man/create_CRU_stack.Rd
@@ -91,4 +91,3 @@ CRU_pre_tmp <- create_CRU_stack(pre = TRUE, tmp = TRUE)
 \seealso{
 \code{\link{create_CRU_df}}
 }
-
