### 2017-02-01 d9a2707
Remove tests to shorten testing time
Will need to do further work like what was done for elevation to
increase code coverage
