### 2017-02-01 8d14271
Update to 0.1.3 for CRAN release
Update DESCRIPTION, documentation, NEWS.md and cran-comments.md
