### 2017-02-01 b5c4949
Update cran-comments
Travis seems to be broken at the moment, so no Travis logs for this bug
fix release
