### 2017-02-01 b07cffa
Update tests to run faster
Create own data files on the fly for unit tests so that the tests run
faster
