+++ tests/testthat/test-get_CRU.R
@@ -1,13 +1,15 @@
 context(".get_CRU")
 
-unlink(list.files(path = tempdir(), pattern = ".dat.gz$", full.names = TRUE))
+unlink(list.files(
+  path = tempdir(),
+  pattern = ".dat.gz$",
+  full.names = TRUE
+))
 
 test_that("get_CRU will retrieve only precipitation file when pre TRUE", {
   skip_on_cran()
-  skip_on_appveyor()
-  skip_on_travis()
 
-    # be sure to start with a clean directory
+  # be sure to start with a clean directory
   cache_dir <- tempdir()
   unlink(list.files(
     path = cache_dir,
@@ -15,22 +17,28 @@ test_that("get_CRU will retrieve only precipitation file when pre TRUE", {
     full.names = TRUE
   ))
 
-  .get_CRU(pre = TRUE,
-           pre_cv = FALSE,
-           rd0 = FALSE,
-           tmp = FALSE,
-           dtr = FALSE,
-           reh = FALSE,
-           tmn = FALSE,
-           tmx = FALSE,
-           sunp = FALSE,
-           frs = FALSE,
-           wnd = FALSE,
-           elv = FALSE,
-           cache_dir)
+  .get_CRU(
+    pre = TRUE,
+    pre_cv = FALSE,
+    rd0 = FALSE,
+    tmp = FALSE,
+    dtr = FALSE,
+    reh = FALSE,
+    tmn = FALSE,
+    tmx = FALSE,
+    sunp = FALSE,
+    frs = FALSE,
+    wnd = FALSE,
+    elv = FALSE,
+    cache_dir
+  )
   files <- list.files(cache_dir, pattern = ".dat.gz$")
   expect_identical(files, "grid_10min_pre.dat.gz")
-  unlink(list.files(path = cache_dir, pattern = ".dat.gz$", full.names = TRUE))
+  unlink(list.files(
+    path = cache_dir,
+    pattern = ".dat.gz$",
+    full.names = TRUE
+  ))
 })
 
 
@@ -39,92 +47,57 @@ test_that("get_CRU will retrieve only precipitation file when pre_cv TRUE", {
   skip_on_appveyor()
   skip_on_travis()
   cache_dir <- tempdir()
-  .get_CRU(pre = FALSE,
-           pre_cv = TRUE,
-           rd0 = FALSE,
-           tmp = FALSE,
-           dtr = FALSE,
-           reh = FALSE,
-           tmn = FALSE,
-           tmx = FALSE,
-           sunp = FALSE,
-           frs = FALSE,
-           wnd = FALSE,
-           elv = FALSE,
-           cache_dir)
+  .get_CRU(
+    pre = FALSE,
+    pre_cv = TRUE,
+    rd0 = FALSE,
+    tmp = FALSE,
+    dtr = FALSE,
+    reh = FALSE,
+    tmn = FALSE,
+    tmx = FALSE,
+    sunp = FALSE,
+    frs = FALSE,
+    wnd = FALSE,
+    elv = FALSE,
+    cache_dir
+  )
   files <- list.files(cache_dir, pattern = ".dat.gz$")
   expect_identical(files, "grid_10min_pre.dat.gz")
-  unlink(list.files(path = cache_dir, pattern = ".dat.gz$", full.names = TRUE))
+  unlink(list.files(
+    path = cache_dir,
+    pattern = ".dat.gz$",
+    full.names = TRUE
+  ))
 })
-
-test_that("get_CRU will retrieve only tmp file when tmp TRUE", {
-  skip_on_cran()
-  skip_on_appveyor()
-  skip_on_travis()
-  cache_dir <- tempdir()
-  .get_CRU(pre = FALSE,
-           pre_cv = FALSE,
-           rd0 = FALSE,
[…]

[…]
