### 2017-02-01 #96 @sckott
reviewers assigned, see https://github.com/ropensci/onboarding/issues/96#issuecomment-276509383
