### 2017-02-01 3e44b59
Remove test for downloading precip, check only pre_cv
