### 2017-02-01 23258a3
Fix bugs in and enhance tests for create_CRU_df
