### 2017-02-03 25c2a6a
Fix bugs related to elevation
If elevation was selected along with any other variable the function
would stop and error.

Also, lat/lon were incorrectly reported in the elv data frame
