### 2017-02-03 e96ce7d
Fix bugs introduced fixing bugs related to elevation data in dataframe output
