### 2017-02-03 f2c0636
Fix one more bug related to elevation in the tidy df
