+++ R/tidy_df.R
@@ -36,10 +36,10 @@
     CRU_df <- Reduce(function(...)
       dplyr::full_join(..., by = c("lat", "lon", "month")), CRU_list)
     CRU_df <- dplyr::left_join(CRU_df, elv_df, by = c("lat", "lon"))
-  } else
+  } else {
     CRU_df <- as.data.frame(CRU_list["elv"])
-  names(CRU_df) <- c("lat", "lon", "elv")
-
+    names(CRU_df) <- c("lat", "lon", "elv")
+  }
   return(CRU_df)
 
   # cleanup before we go -----------------------------------------------------
