### 2017-02-03 c845ccb
Update test and one more bugfix for elv
