### 2017-02-03 e6b461c
One more fix in the tests
