+++ cran-comments.md
@@ -8,9 +8,7 @@
 0 errors | 0 warnings | 1 note
 
 * This is minor release with the following changes
-  * Correct fix bug in raster object generation where the objects were incorrectly cropped  
-  * Update documentation with ROxygen 6.0.0  
-  * Minor edits to documentation for clarity  
+  * Correct fix bug in data frame object generation where elevation was improperly handled and function would stop  
   
 * Check logs from local OSX, Appveyor and win-builder are all available from https://github.com/adamhsparks/checkresults/tree/master/getCRUCLdata
 
