### 2017-02-03 5cace46
Update tests to test .get_CRU on travis
