### 2017-02-03 a702d8d
Add test for tmn/tmx calculations
