### 2017-02-03 c7e3a1d
Format tests file
