+++ tests/testthat/test-calculate_tmn_tmx.R
@@ -1,10 +1,10 @@
 
 test_that("tmx is calculated correctly", {
-tmx <- .calculate_tmx(10.0, 2.0)
-expect_equal(tmx, 11)
+  tmx <- .calculate_tmx(10.0, 2.0)
+  expect_equal(tmx, 11)
 })
 
 test_that("tmn is calculated correctly", {
-tmn <- .calculate_tmn(10.0, 2.0)
-expect_equal(tmn, 9)
+  tmn <- .calculate_tmn(10.0, 2.0)
+  expect_equal(tmn, 9)
 })
