### 2017-02-17 #96 @ldecicco-USGS
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

I'm not sure if this is actually an issue with the package, but when I tried to install via CRAN, there was an error via the Windows binary:

```
> install.packages("getCRUCLdata")
Installing package into ‘C:/Users/ldecicco/Documents/R/win-library/3.3’
(as ‘lib’ is unspecified)
trying URL 'https://cloud.r-project.org/bin/windows/contrib/3.3/getCRUCLdata_0.1.4.zip'
Content type 'application/zip' length 160600 bytes (156 KB)
downloaded 0 bytes

Warning in install.packages :
  downloaded length 0 != reported length 160600
Warning in install.packages :
  error 1 in extracting from zip file
Warning in install.packages :
  cannot open compressed file 'getCRUCLdata/DESCRIPTION', probable reason 'No such file or directory'
Error in install.packages : cannot open the connection
```

I could install via source just fine. This could be an issue with my CRAN mirror or something else...but I haven't experienced it before.

#### Vignette

The first thing I tried to do was run the first example in the vignette, and was unsuccessful:

```
library(getCRUCLdata)
CRU_data <- create_CRU_df(pre = TRUE,
                          pre_cv = TRUE,
                          rd0 = TRUE,
                          tmp = TRUE,
                          dtr = TRUE,
                          reh = TRUE,
                          tmn = TRUE,
                          tmx = TRUE,
                          sunp = TRUE,
                          frs = TRUE,
                          wnd = TRUE,
                          elv = TRUE)

Downloading requested data files.

Error in .f(.x[[i]], ...) : Too many retries...server may be under load
```

I then tried setting a single parameter to TRUE, and still could not get data to return:

```
CRU_data <- create_CRU_df(pre = TRUE)

Downloading requested data files.

Error in .f(.x[[i]], ...) : Too many retries...server may be under load
```

#### Examples

I had the same problem running the examples from the help files as running the examples from the vignette.

Similarly, the tests via `testthat` didn't pass with the same error "Too many retries...server may be under load". However, looking at the tests, they looked like appropriate tests. According to Coveralls, there's a 59% coverage on code. This isn't great, but it's not too bad.

#### Troubleshooting

When I went to the URLs to test the service, they worked fine. I use `httr` for the package `dataRetrieval` (so, not as familiar with `purr`). I tried this:

```
library(httr)

temp <- tempfile()
temp <- paste0(temp,".dat.gz")
obs_url <- "https://crudata.uea.ac.uk/cru/data/hrg/tmc/grid_10min_dtr.dat.gz"
doc <- GET(obs_url, write_disk(temp))

headerInfo <- headers(doc)
doc2 <- read.table(gzfile(temp))

unlink(doc)
```

`doc2` came back as a table with 14 columns and 566262 rows. So, maybe give `httr` a try? As it is, I can't get anything back from the package, and I don't think it's the underlying services fault.
