### 2017-02-17 #96 @adamhsparks
Thanks for the review @ldecicco-USGS.

It's interesting. I'm not able to replicate the issue with installation because unfortunately, I don't have a Windows machine to test on, I rely on Winbuilder and Appveyor to alert me to any issues. I checked the zip package from CRAN, if I'm reading the error correctly it says that there was no DESCRIPTION file. However, the zip file I downloaded had one in it.

Moving on to the other items, I'm also unable to replicate the issues. I just ran the code that was pasted here that timed out. It worked and the tests pass on my computer, Travis and Appveyor when downloading, so I assume everything is fine. I wonder if there's a network issue that is causing problems?

I'll look at using `httr`. The `purr` code comes from one of my other packages where the server will stop responding so it limits the number of requests made and I incorporated it here to be nice to this server as well, so I'm stumped about what the issue is.
