### 2017-02-18 c5d604d
Change method used to download files
Use `httr::GET()` to download files after review from @ldecicco-USGS
wherein issues were identified using `curl` and `purrr` method.
