+++ man/create_CRU_df.Rd
@@ -88,4 +88,3 @@ CRU_pre_tmp <- create_CRU_df(pre = TRUE, tmp = TRUE)
 \seealso{
 \code{\link{create_CRU_stack}}
 }
-
