+++ DESCRIPTION
@@ -17,15 +17,15 @@ License: MIT + file LICENSE
 Depends:
     R (>= 3.0.0)
 Imports:
-    curl,
     dplyr,
+    httr,
     plyr,
     purrr (>= 0.2.0),
     raster,
     tidyr,
     utils
 LazyData: TRUE
-RoxygenNote: 5.0.1
+RoxygenNote: 6.0.1
 ByteCompile: TRUE
 Suggests:
     testthat,
