### 2017-02-19 d1fca9d
Return data frame as a tibble, use progress bar for download
