+++ DESCRIPTION
@@ -6,10 +6,10 @@ Authors@R: person(given = "Adam", family = "Sparks", email = "adamhsparks@gmail.
 Description: Provides functions that automate downloading and importing
     University of East Anglia Climate Research Unit (CRU) CL2.0 climatology data
     into R, facilitates the calculation of minimum temperature and maximum
-    temperature and formats the data into a tidy data frame or a list of raster
-    stack objects for use in an R session.  CRU CL2.0 data are a gridded
-    climatology of 1961-1990 monthly means released in 2002 and cover all land
-    areas (excluding Antarctica) at 10-minute resolution.  For more
+    temperature and formats the data into a tidy data frame as a tibble or a 
+    list of raster stack objects for use in an R session.  CRU CL2.0 data are a
+    gridded climatology of 1961-1990 monthly means released in 2002 and cover
+    all land areas (excluding Antarctica) at 10-minute resolution.  For more
     information see the description of the data provided by the University of
     East Anglia Climate Research Unit,
     <https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt>.
@@ -23,6 +23,7 @@ Imports:
     purrr (>= 0.2.0),
     raster,
     tidyr,
+    tibble,
     utils
 LazyData: TRUE
 RoxygenNote: 6.0.1
