### 2017-02-19 #96 @adamhsparks
@ldecicco-USGS, I've modified the download function to use `httr::GET` as suggested in the devel branch, if you'd like to check it.

https://github.com/adamhsparks/getCRUCLdata/tree/devel
