### 2017-02-21 #96 @ivanhanigan
## Package Review for getCRUCLdata

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

- 3hrs

---

This is a useful R package that is well written and provides access to an important data collection.
I reviewed the Github version, master branch, installed today 21 Feb 2017, on Ubuntu linux.  I made comments against the line numbers for files in the devel branch.
The documentation is clear and attractive.  Functionality on my linux machine is mostly good, with some functionality not working/returning errors.

#### Major comments

##### test 1 `create_CRU_df`

- Success: I ran the suggested line from README for all 12 parameters and retrieved a df with 6800028 obs, 15 vars, 0.8GB according to object.size()
- Comment re single threaded.  I ran this on a 16 core machine and was disappointed to only see one core spinning.  Is there options for parallelisation?
- Comment re size of download is significant, and any problems (e.g. broken connection with cloud VM, power outage etc) may cause frustration if object is lost due to crashed R session.  I ran save.image() straight after successful
- Issue re `with(CRU_data, plot(lon, lat))` caused my Rstudio server to return status 502 (unresponsive) and I killed the session.  I tried again, left it running for 30 mins and then logged in to see the plot had completed, but the session was still clunky and I got unresponsive warnings and then I terminated it.  This is a NeCTAR research cloud XXL running on QCIF infrastructure (64GB, 16 core, 2.3 GHz Opterons) so should not have an issue.  Suggest you pack some easily visualisable meta info such as the lat lon points as a small df or spatialPoints obj so is easily plotable.
- Issue re next line causes a warning and fails

t_rh <- create_CRU_df(tmp = TRUE,
+                       reh = TRUE)

Downloading requested data files.

  |=================================================================================================================| 100%
Error in full_join_impl(x, y, by$x, by$y, suffix$x, suffix$y) :
  'lat' column not found in lhs, cannot join
```

- Issue re help file for this function, has example but also fails

 CRU_pre_tmp <- create_CRU_df(pre = TRUE, tmp = TRUE)

Downloading requested data files.

  |=================================================================================================================| 100%
Error in full_join_impl(x, y, by$x, by$y, suffix$x, suffix$y) :
  'lat' column not found in lhs, cannot join
```

##### test 2 `create_CRU_stack`

- Comment re I then went ahead and test create_CRU_stack.  I immediatly realise this is downloading more data, and wondered if seeing as I have a dataframe I previously downloaded it might be possible to just convert that to a stack. To save more downloading.  Converse is also true: if user has already downloaded raster stack (or stored this as local tifs) then perhaps create_CRU_df could be told to use the data that already exists.
- Comment re I see in devel/vignettes/getCRUCLdata.Rmd line 112-115 the useful suggestion to store rasters locally as geotiffs.  Recommend a similar suggestion above for the dataframe.  Also suggest you load the raster library here too, for Noobs who may get upset by the warning if the try to run writeRaster without that package attached.
- Success re CRU_stack.  I ran the following and got successful geotiffs

writeRaster(CRU_stack$pre, filename = paste0("~/tmp/pre_", names(CRU_stack$pre)), bylayer = TRUE, format = "GTiff")
writeRaster(CRU_stack$reh, filename = paste0("~/tmp/reh_", names(CRU_stack$reh)), bylayer = TRUE, format = "GTiff")
```

- loaded precip december tif into QGIS succeeds, and is projected as WGS84, thanks.  Value for Darwin of 240 is very close to the Darwin Airport 233.9 reported by the BoM for the same 61-90 period.  Awesome thanks.
- issue re next line from vignettes/getCRUCLdata

tmn_tmx <- create_CRU_stack(tmn = TRUE,
                            tmx = TRUE)
Downloading requested data files.

  |======================================                                                                           |  33%
Error in raster::crop(y, raster::extent(-180, 180, -60, 85)) :
  object 'y' not found

```
- Issue re help for this function: "Details This function generates a data.frame object in R with the following possible fields as specified by the user: " should read "raster stack" not data.frame
- Success the example in that help file worked (`CRU_pre_tmp <- create_CRU_stack(pre = TRUE, tmp = TRUE)`)

#### Minor comments

Re devel/README.Rmd:

- line 34 re '10 minute resolution' - can you say this in decimal degrees to avoid confusion for GIS Noobs (also in devel/vignettes/getCRUCLdata.Rmd, line 19).
- further comment, this appears in the file `devel/man/getCRUdata.Rd` as 10 arc-second, so at least need to resolve which units to declare, but I strongly recommend aligning with decimal degrees.
- line 57 - should read "stable version of getCRUCLdata is available from CRAN" not "GSODR is available"

Re paper.md

- This could have more of a statement of need and audience.  Currently it echos the README but I think here it would be nice to expand a bit of discussion.

Re general

- There might be opportunity to optimise the processing and also provide some safegaurds in terms of automatically storing downloaded data locally to protect users from crashs etc that might lead to them having to download the same data multiple times.
- I did not assess code style, code duplication, Automated tests or Packaging guidelines as I lack the programming expertise for that.
- I checked a couple external URLs which succeed in taking me to the Uni East Anglia pages that describe these data, thanks!

#### My sessionInfo

Running under: Ubuntu 15.10

 [1] LC_CTYPE=en_AU.UTF-8       LC_NUMERIC=C               LC_TIME=en_AU.UTF-8        LC_COLLATE=en_AU.UTF-8
 [5] LC_MONETARY=en_AU.UTF-8    LC_MESSAGES=en_AU.UTF-8    LC_PAPER=en_AU.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=en_AU.UTF-8 LC_IDENTIFICATION=C

[1] getCRUCLdata_0.1.4

[1] magrittr_1.5   R6_2.2.0       assertthat_0.1 DBI_0.5-1      tools_3.3.1    dplyr_0.5.0    tibble_1.2     Rcpp_0.12.9
```
