### 2017-02-21 0423075
Fix documentation and README errors
thanks to @ivanhanigan
