+++ README.Rmd
@@ -54,7 +54,7 @@ Install
 
 ## Stable version
 
-A stable version of GSODR is available from
+A stable version of getCRUCLdata is available from
 [CRAN](https://cran.r-project.org/package=getCRUCLdata).
 
 ``` r
