### 2017-02-21 843534a
Revise resolution description
@ivanhanigan suggested listing resolution as degrees. I've included
this measurement of resolution along with the original 10 arcminute
description. However, New et al. only refer to 10 "minute" data, I
refer to it as 10 "arcminute".
