+++ man/getCRUdata.Rd
@@ -12,9 +12,9 @@ minimum temperature and maximum temperature, converts elevation from
 kilometres to metres and formats the data into a tidy data frame or a list
 of raster stack objects for use in an R session.  CRU CL2.0 data are a
 gridded climatology of 1961-1990 monthly means released in 2002 and cover all
-land areas (excluding Antarctica) at 10 arc-minute resolution.  For more
-information see the description of the data provided by the University of
-East Anglia Climate Research Unit (CRU),
+land areas (excluding Antarctica) at 10 arcminutes (0.1666667 degree)
+resolution.  For more information see the description of the data provided
+by the University of East Anglia Climate Research Unit (CRU),
 \url{https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt}.
 }
 \references{
