+++ README.Rmd
@@ -31,9 +31,9 @@ importing CRU CL2.0 climatology data into R, facilitate the calculation of
 minimum temperature and maximum temperature, and formats the data into a tidy
 data frame or a list of raster stack objects for use in an R session. CRU CL2.0
 data are a gridded climatology of 1961-1990 monthly means released in 2002 and
-cover all land areas (excluding Antarctica) at 10-minute resolution. For
-more information see the description of the data provided by the University of
-East Anglia Climate Research Unit (CRU),
+cover all land areas (excluding Antarctica) at 10 arcminutes (0.1666667 degree)
+resolution. For more information see the description of the data provided by the
+University of East Anglia Climate Research Unit (CRU),
 <https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt>.
 
 ## Changes to original data
