### 2017-02-21 #96 @adamhsparks
@ivanhanigan Thank you for a thorough and helpful review.

I've started addressing a few of the more minor issues that you've pointed out in the "devel" branch and will continue to work through these suggestions.
