### 2017-02-21 7f5ec14
Use read_fwf to import data faster
