### 2017-02-21 948953b
Replace `read_fwf` with `fread`
Using `data.table::fread` is a drop-in solution to replace the original
`utils::read.table` and is ~5X faster
