+++ DESCRIPTION
@@ -17,6 +17,7 @@ License: MIT + file LICENSE
 Depends:
     R (>= 3.0.0)
 Imports:
+    data.table,
     dplyr,
     httr,
     plyr,
