### 2017-02-21 7b0fc7a
convert `data.table` to `data.frame` when importing
