### 2017-02-21 5739daa
Update documentation
Help file for `getCRUCLdata` did not exist
