+++ R/getCRUCLdata.R
@@ -1,6 +1,6 @@
-#' getCRUdata: Download and Create a Data Frame or Raster Stack Object of CRU CL2.0 Climatology Variables
+#' getCRUCLdata: Download and Create a Data Frame or Raster Stack Object of CRU CL2.0 Climatology Variables
 #'
-#' The getCRUdata package provides functions that automate downloading and
+#' The getCRUCLdata package provides functions that automate downloading and
 #' importing CRU CL2.0 climatology data into R, facilitates the calculation of
 #' minimum temperature and maximum temperature, converts elevation from
 #' kilometres to metres and formats the data into a tidy data frame or a list
@@ -13,7 +13,7 @@
 #'
 #' @docType package
 #'
-#' @name getCRUdata
+#' @name getCRUCLdata
 #'
 #' @seealso \code{\link{create_CRU_df}}
 #'
