+++ man/getCRUCLdata.Rd
@@ -1,12 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/getCRUCLdata.R
 \docType{package}
-\name{getCRUdata}
-\alias{getCRUdata}
-\alias{getCRUdata-package}
-\title{getCRUdata: Download and Create a Data Frame or Raster Stack Object of CRU CL2.0 Climatology Variables}
+\name{getCRUCLdata}
+\alias{getCRUCLdata}
+\alias{getCRUCLdata-package}
+\title{getCRUCLdata: Download and Create a Data Frame or Raster Stack Object of CRU CL2.0 Climatology Variables}
 \description{
-The getCRUdata package provides functions that automate downloading and
+The getCRUCLdata package provides functions that automate downloading and
 importing CRU CL2.0 climatology data into R, facilitates the calculation of
 minimum temperature and maximum temperature, converts elevation from
 kilometres to metres and formats the data into a tidy data frame or a list
