### 2017-02-22 ec17d45
Used cached files already downloaded
If data files have already been downloaded during current R session,
the package will now check the `tempdir` to see if the files exist
before downloading to save time. For example, a user creates a tidy
data frame of temperature data and decides to create a raster stack of
temperature and relative humidity while running the same R session. The
package will only download relative humidity and reuse the temperature
data file.
