+++ R/get_CRU.R
@@ -64,19 +64,33 @@
 
     files <- as.list(files[object_list %in% !isTRUE(files)])
 
-    # download files -----------------------------------------------------------
-    message(" \nDownloading requested data files.\n ")
-    pb <-
-      utils::txtProgressBar(
-        min = 0,
-        max = length(files),
-        initial = 0,
-        style = 3
-      )
-    for (f in 1:length(files)) {
-      httr::GET(url = files[[f]], httr::write_disk(paste0(cache_dir, "/",
-                                                 basename(files[[f]])),
-                                          overwrite = TRUE))
-      utils::setTxtProgressBar(pb, f)
+    # check to see, did we already download these data, if so don't redownload--
+    cache_dir_contents <-
+      list.files(cache_dir, pattern = ".dat.gz$")
+    cache_dir_contents <- paste0(CRU_url, cache_dir_contents)
+
+    # if there are files already locally available;
+    # check against the newly requested files and update list to download
+    if (length(cache_dir_contents) > 0) {
+      files <- as.list(files[!(files %in% cache_dir_contents)])
+    }
+
+    if (length(files) > 0) {
+      # download files ---------------------------------------------------------
+      message(" \nDownloading requested data files.\n ")
+      pb <-
+        utils::txtProgressBar(
+          min = 0,
+          max = length(files),
+          initial = 0,
+          style = 3
+        )
+      for (f in 1:length(files)) {
+        httr::GET(url = files[[f]],
+                  httr::write_disk(paste0(cache_dir, "/",
+                                          basename(files[[f]])),
+                                   overwrite = TRUE))
+        utils::setTxtProgressBar(pb, f)
+      }
     }
   }
