### 2017-02-22 349a320
Additions to vignette documentation
Add example of writing CSV file to disk.

Load `raster` package before calling `writeRaster()`
