### 2017-02-22 912e485
Check local files before stack/df creation
Check files in `cache_dir`, remove files that are not requested to be
in resulting stack or data frame.
