### 2017-02-22 ea2876e
Improved documentation and sort data frame by month
Improved README.Rmd documentation, showing how to plot stacks and data
frames.

Added capabilities to vignette.

TODO: add graphs to vignette.
