+++ R/create_CRU_df.R
@@ -126,5 +126,25 @@ create_CRU_df <- function(pre = FALSE,
       CRU_df <- subset(CRU_df, select = -dtr)
     }
   }
+  CRU_df$month <- factor(
+    CRU_df$month,
+    levels <-  c(
+      "jan",
+      "feb",
+      "mar",
+      "apr",
+      "may",
+      "jun",
+      "jul",
+      "aug",
+      "sep",
+      "oct",
+      "nov",
+      "dec"
+    )
+  )
+
+  CRU_df <- dplyr::arrange(CRU_df, month)
+
   return(tibble::as_tibble(CRU_df))
 }
