### 2017-02-22 530dce6
Add images to vignette and other documentation updates
A few minor fixes along with the images now being added to the vignette
