### 2017-02-23 b514593
Update graphs and maps, update documentation
Update README.md and vignette

Update corresponding docs/ pages.
