### 2017-02-23 a406c70
Fix documentation error
create_CRU_stack documentation incorrectly stated in details that it
created a data.frame object, it should state that it creates a raster
stack object.
