### 2017-02-23 c2bfbe6
Add "Suggests" field to DESCRIPTION and fix tests
Add `ggplot2` and `viridis` to Suggests field.

Fix tests

Fix README.Rmd, don't evaluate all R chunks

Create `month` object in `create_CRU_df()` to fill later.
