+++ README.md
@@ -44,7 +44,7 @@ if (!require("devtools")) {
   install.packages("devtools")
 }
 
-#devtools::install_github("adamhsparks/getCRUCLdata", build_vignettes = TRUE)
+devtools::install_github("adamhsparks/getCRUCLdata", build_vignettes = TRUE)
 
 library(getCRUCLdata)
 ```
