### 2017-02-23 86b765f
Fix documentation issues
Clean up header formats and clarify documentation
