### 2017-02-23 6d0087b
Fix DESCRIPTION file
No need for two "Suggests" fields
