### 2017-02-23 18e3b4f
Add `skip_on_cran` and `skip_on_appveyor` to tests
Some tests take a bit of time to download. Skip these tests on cran and
appveyor.
