### 2017-02-23 0f24f4e
Add `skip_on_travis()` to tests
