### 2017-02-23 84718ae
Update test environments in cran-comments.md
