+++ cran-comments.md
@@ -2,15 +2,13 @@
 * local macOS install, R 3.3.2
 * Windows (on Appveyor), R 3.3.2
 * win-builder (devel and release)
+* Linux (on Travis), (devel and release)
+* local Linux (Ubuntu 16.04) install, R 3.3.2
 
 ## R CMD check results
 
 0 errors | 0 warnings | 1 note
 
-* This is minor release with the following changes
-  * Correct fix bug in data frame object generation where elevation was improperly handled and function would stop  
-  
-* Check logs from local OSX, Appveyor and win-builder are all available from https://github.com/adamhsparks/checkresults/tree/master/getCRUCLdata
 
 ## Reverse dependencies
 
