### 2017-02-23 d9accb7
Changing tests to run more quickly
