+++ tests/testthat/test-create_CRU_df.R
@@ -12,33 +12,28 @@ test_that("create_CRU_df fails if no parameters are TRUE", {
                "You must select at least one parameter for download.")
 })
 
-test_that("Test that .tidy_df creates a tidy dataframe of pre, pre_cv and tmp", {
+test_that("Test that .tidy_df creates a tidy dataframe of pre and pre_cv", {
   skip_on_cran()
   skip_on_appveyor()
-  skip_on_travis()
   CRU_df <-
     create_CRU_df(pre_cv = TRUE,
-                  pre = TRUE,
-                  tmp = TRUE)
+                  pre = TRUE)
 
   expect_true(is.data.frame(CRU_df))
-  expect_named(CRU_df, c("lat", "lon", "month", "pre", "pre_cv", "tmp"))
+  expect_named(CRU_df, c("lat", "lon", "month", "pre", "pre_cv"))
   expect_is(CRU_df$lat, "numeric")
   expect_is(CRU_df$lon, "numeric")
   expect_is(CRU_df$month, "factor")
   expect_is(CRU_df$pre, "numeric")
   expect_is(CRU_df$pre_cv, "numeric")
-  expect_is(CRU_df$tmp, "numeric")
 })
 
-test_that("Test that .tidy_df creates a tidy dataframe of pre, pre_cv, tmp and elv", {
+test_that("Test that .tidy_df creates a tidy dataframe of pre, pre_cv, and elv", {
   skip_on_cran()
   skip_on_appveyor()
-  skip_on_travis()
   CRU_df <-
     create_CRU_df(pre_cv = TRUE,
                   pre = TRUE,
-                  tmp = TRUE,
                   elv = TRUE)
 
   expect_true(is.data.frame(CRU_df))
