### 2017-02-23 f9dece3
Fix tests, again
This time it worked, I swear
