### 2017-02-23 163ca41
still trying to weed out issue with elv stack
