+++ man/create_CRU_df.Rd
@@ -58,7 +58,7 @@ Nomenclature and units from readme.txt:
  \describe{
    \item{cv}{cv of precipitation (percent)}
  }
-\item{rd0}{wet-days (no days with >0.1mm rain per month)}
+\item{rd0}{wet-days (number days with >0.1mm rain per month)}
 \item{tmp}{mean temperature (degrees Celsius)}
 \item{dtr}{mean diurnal temperature range (degrees Celsius)}
 \item{reh}{relative humidity (percent)}
