### 2017-02-24 18331f4
Split get and create functions for stack
user can now create stacks with list of locally available files
