+++ R/create_stacks.R
@@ -1,0 +1,134 @@
+#' @noRd
+#'
+create_stacks <- function(files, tmn, tmx, tmp, dtr, pre, pre_cv) {
+  message("\nCreating raster stack now.\n")
+  wrld <-
+    raster::raster(
+      nrows = 930,
+      ncols = 2160,
+      ymn = -65,
+      ymx = 90,
+      xmn = -180,
+      xmx = 180
+    )
+
+  wrld[] <- NA
+
+  month_names <-
+    c("jan",
+      "feb",
+      "mar",
+      "apr",
+      "may",
+      "jun",
+      "jul",
+      "aug",
+      "sep",
+      "oct",
+      "nov",
+      "dec")
+
+  # Create raster objects using cellFromXY and generate a raster stack
+  # create.stack takes pre, tmp, tmn and tmx and creates a raster
+  # object stack of 12 month data
+
+  files <-
+    list.files(tempdir(), pattern = ".dat.gz$", full.names = TRUE)
+
+  CRU_stack_list <-
+    plyr::llply(.fun = .create_stack,
+                files,
+                wrld,
+                month_names,
+                .pre = pre,
+                .pre_cv = pre_cv,
+                .progress = "text")
+
+  names(CRU_stack_list) <- substr(basename(files), 12, 14)
+
+  # cacluate tmn -------------------------------------------------------------
+  if (isTRUE(tmn)) {
+    CRU_stack_list$tmn <-
+      CRU_stack_list$tmp - (0.5 * CRU_stack_list$dtr)
+  }
+  # cacluate tmx -------------------------------------------------------------
+  if (isTRUE(tmx)) {
+    CRU_stack_list$tmx <-
+      CRU_stack_list$tmp + (0.5 * CRU_stack_list$dtr)
+  }
+
+  # cleanup if tmn/tmx specified but tmp/dtr not -----------------------------
+  if (!isTRUE(tmp) | !isTRUE(dtr) & isTRUE(tmx) | isTRUE(tmn)) {
+    CRU_stack_list[which(names(CRU_stack_list) %in% c("tmp", "dtr"))] <-
+      NULL
+  }
+  if (!isTRUE(dtr) & isTRUE(tmx) | isTRUE(tmn)) {
+    CRU_stack_list[which(names(CRU_stack_list) %in% "dtr")] <- NULL
+  }
+  if (!isTRUE(tmp) & isTRUE(tmx) | isTRUE(tmn)) {
+    CRU_stack_list[which(names(CRU_stack_list) %in% "tmp")] <- NULL
+  }
+  return(CRU_stack_list)
+}
+
+#' @noRd
+.create_stack <- function(files,
+                          wrld,
+                          month_names,
+                          .pre,
+                          .pre_cv) {
+  wvar <-
+    data.frame(data.table::fread(paste0("gzip -dc ", files),
+                                 header = FALSE))
+  cells <- raster::cellFromXY(wrld, wvar[, c(2, 1)])
+  if (ncol(wvar) == 14) {
+    for (j in 3:14) {
+      wrld[cells] <- wvar[, j]
+      if (j == 3) {
+        y <- wrld
+      } else
+        y <- raster::stack(y, wrld)
+    }
+    names(y) <- month_names
+  } else if (ncol(wvar) == 26) {
+    if (isTRUE(.pre) & isTRUE(.pre_cv)) {
+      for (k in 3:26) {
+        wrld[cells] <- wvar[, k]
+        if (k == 3) {
+          y <- wrld
+        } else
+          y <- raster::stack(y, wrld)
+      }
+      names(y) <- c(month_names, paste0(".pre_cv_", month_names))
+    } else if (isTRUE(.pre)) {
+      for (k in 3:14) {
+        wrld[cells] <- wvar[, k]
[…]
