+++ R/create_CRU_stack.R
@@ -1,8 +1,8 @@
-#' @title Download and Create a List of Raster Stack Objects From CRU CL2.0 Climatology Variables
+#' @title Create a List of Raster Stack Objects From CRU CL2.0 Climatology Variables on Local Disk
 #'
-#'@description This function automates downloading and importing CRU CL2.0
-#'climatology data into R and creates a list of raster stacks of the data.  If
-#'requested, minimum and maximum temperature may also be automatically
+#'@description This function automates importing CRU CL2.0 climatology data into
+#'R from locally available data files and creates a list of raster stacks of the
+#'data.  If requested, minimum and maximum temperature may also be automatically
 #'calculated as described in the data readme.txt file.
 #'
 #'Nomenclature and units from readme.txt:
@@ -23,40 +23,32 @@
 #'For more information see the description of the data provided by CRU,
 #'\url{https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt}
 #'
-#' @details This function generates a raster stack object in R with the following
-#' possible fields as specified by the user:
-#' @param pre Logical. Fetch precipitation (millimetres/month) from server and
-#' return in a raster stack? Defaults to FALSE.
-#' @param pre_cv Logical. Fetch cv of precipitation (percent) from server and
-#' return in a raster stack? Defaults to FALSE.
-#' @param rd0 Logical. Fetch wet-days (number days with >0.1 millimetres rain
-#' per month) and return in a raster stack? Defaults to FALSE.
-#' @param dtr Logical. Fetch mean diurnal temperature range (degrees Celsius)
-#' and return it in a raster stack? Defaults to FALSE.
-#' @param tmp Logical. Fetch temperature (degrees Celsius) and return it in the
-#' raster stack? Defaults to FALSE.
+#' @details This function generates a raster stack object in R from local files
+#' already downloaded via a web browser or FTP. The user will need to specify
+#' the following options.
+#' @param pre_cv Logical. Create a raster stack of precipitation CV? Requires
+#' precipitation file to be available. Defaults to FALSE.
 #' @param tmn Logical. Calculate minimum temperature values (degrees Celsius)
-#' and return it in a raster stack? Defaults to FALSE.
+#' and return it in a raster stack? Requires tmp and dtr files to both be
+#' available for calculation. Defaults to FALSE.
 #' @param tmx Logical. Calculate maximum temperature (degrees Celsius) and
-#' return it in a raster stack? Defaults to FALSE.
-#' @param reh Logical. Fetch relative humidity and return it in a raster stack?
-#' Defaults to FALSE.
-#' @param sunp Logical. Fetch sunshine, percent of maximum possible (percent of
-#' day length) and return it in raster stack? Defaults to FALSE.
-#' @param frs Logical. Fetch ground-frost records (number of days with ground-
-#' frost per month) and return it in raster stack? Defaults to FALSE.
-#' @param wnd Logical. Fetch 10m wind speed (metres/second) and return it in the
-#' raster stack? Defaults to FALSE.
-#' @param elv Logical. Fetch elevation (converted to metres) and return it in a
-#' raster layer object? Defaults to FALSE.
+#' return it in a raster stack?  Requires tmp and dtr files to both be
[…]
