+++ NAMESPACE
@@ -2,4 +2,5 @@
 
 export(create_CRU_df)
 export(create_CRU_stack)
+export(get_CRU_stack)
 importFrom(dplyr,"%>%")

[…]
