### 2017-02-26 3d82af9
New functions to work with local files and updates to documentation
Add `create_*` functions for working with local data rather than
downloading through R.

Update documentation in vignette and help files.
