+++ NAMESPACE
@@ -2,5 +2,6 @@
 
 export(create_CRU_df)
 export(create_CRU_stack)
+export(get_CRU_df)
 export(get_CRU_stack)
 importFrom(dplyr,"%>%")

[…]
