### 2017-02-26 33fbc9b
Corrections and clarifications in documentation
