+++ README.Rmd
@@ -26,14 +26,14 @@ knitr::opts_chunk$set(
 
 Author/Maintainer: Adam Sparks
 
-The getCRUCLdata package provides two functions that automate downloading and
-importing CRU CL2.0 climatology data into R, facilitate the calculation of
-minimum temperature and maximum temperature, and formats the data into a tidy
-data frame or a list of raster stack objects for use in an R session. CRU CL2.0
-data are a gridded climatology of 1961-1990 monthly means released in 2002 and
-cover all land areas (excluding Antarctica) at 10 arcminutes (0.1666667 degree)
-resolution. For more information see the description of the data provided by the
-University of East Anglia Climate Research Unit (CRU),
+The getCRUCLdata package provides four functions that automate importing CRU 
+CL2.0 climatology data into R, facilitate the calculation of minimum temperature
+and maximum temperature, and formats the data into a tidy data frame or a list
+of raster stack objects for use in an R session. CRU CL2.0 data are a gridded
+climatology of 1961-1990 monthly means released in 2002 and cover all land areas
+(excluding Antarctica) at 10 arcminutes (0.1666667 degree) resolution. For more
+information see the description of the data provided by the University of East
+Anglia Climate Research Unit (CRU),
 <https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt>.
 
 ## Changes to original data
@@ -43,7 +43,6 @@ This package automatically converts elevation values from kilometres to metres.
 This package crops all spatial outputs to an extent of ymin = -60, ymax = 85,
 xmin = -180, xmax = 180. Note that the original wind data include land area for
 parts of Antarctica.
-    
 
 
 Quick Start
@@ -82,30 +81,39 @@ Using getCRUCLdata
 
 ## Creating tidy data frames for use in R
 
-The `create_CRU_df()` function returns a tidy data frames, as a [`tibble`](https://github.com/tidyverse/tibble) object, of the CRU CL2.0 climatology elements. Illustrated here, create a tidy data frame of all CRU CL2.0 climatology elements available.
+There are two methods that this package provides for creating tidy data frames of the CRU CL2.0 climate data as a [`tibble`](https://github.com/tidyverse/tibble). Depending on your needs you may need to use `create_CRU_df()`, which will create a tidy data frame of the climate elements from *locally available* files. Or you can use `get_CRU_df()`, which will automate the fetching and importing of the data files into R in one step.
 
+The `create_CRU_df()` function is useful if you have network issues that interfere with R downloading the data files themselves from the CRU website or if you frequently work with these data and do not wish to download them every time they are needed.
+
+The `get_CRU_df()` function creates tidy data frames of the CRU CL2.0 climatology elements. Illustrated here, create a tidy data frame of all CRU CL2.0 climatology elements available.
 
 ```{r, eval=FALSE}
-CRU_data <- create_CRU_df(pre = TRUE,
-                          pre_cv = TRUE,
-                          rd0 = TRUE,
-                          tmp = TRUE,
-                          dtr = TRUE,
-                          reh = TRUE,
-                          tmn = TRUE,
-                          tmx = TRUE,
-                          sunp = TRUE,
-                          frs = TRUE,
-                          wnd = TRUE,
-                          elv = TRUE)
+library(getCRUCLdata)
+
+CRU_data <- get_CRU_df(pre = TRUE,
+                       pre_cv = TRUE,
+                       rd0 = TRUE,
+                       tmp = TRUE,
+                       dtr = TRUE,
[…]
