### 2017-02-26 6bb621e
Fix link to viridis on cran, update news and cran-comments
