### 2017-02-26 6bcc564
Add tests for `get_*` functions
