### 2017-02-26 b6cb64a
Add new tests
