### 2017-02-26 642db48
Update paper to include the data that CRU CL2.0 include
