### 2017-02-26 cd7cd61
Minor edits for grammar in paper
