### 2017-02-26 ae02ead
Update cran-comments and NEWS.md
