### 2017-02-26 3709f14
Minor changes in NEWS.md/cran-comments.md
