+++ NEWS.md
@@ -3,7 +3,6 @@
 ## Major Changes
 
   * `create_CRU_stack()` and `create_CRU_df()` now only work with locally available files. If you need to fetch and create a data frame or raster stack of the data, please use the new functions, `get_CRU_stack()` and `create_CRU_stack()`
-  * Months are returned as a factor object in the tidy data frame
   * R >=3.2.0 now required
 
 ## Minor Changes
@@ -12,6 +11,7 @@
   * Change the method in which files are downloaded to use `httr::GET()`
   * Ingest data using `data.table::fread` to decrease the amount of time necessary to run the functions
   * Functions check to see if data file(s) have already been downloaded during current R session, if so data file(s) are not requested for download again
+  * Months are returned as a factor object in the tidy data frame
   
 # getCRUCLdata 0.1.4
 
