### 2017-02-26 #96 @adamhsparks
@ivanhanigan and @ldecicco-USGS,
I've taken your suggestions from the reviews and incorporated them into the latest release, v0.1.5.

# Highlights
## Major Changes

  * `create_CRU_stack()` and `create_CRU_df()` now only work with locally available files. If you need to fetch and create a data frame or raster stack of the data, please use the new functions, `get_CRU_stack()` and `create_CRU_stack()`
  * R >=3.2.0 now required

## Minor Changes

  * Improved documentation with examples on mapping and graphing and more detail regarding the data itself
  * Change the method in which files are downloaded to use `httr::GET()`
  * Ingest data using `data.table::fread` to decrease the amount of time necessary to run the functions
  * Functions check to see if data file(s) have already been downloaded during current R session, if so data file(s) are not requested for download again
  * Months are returned as a factor object in the tidy data frame

# Detailed discussion

@ivanhanigan:

Regarding the parallelisation. I've intentionally not gone this route due to cross-platform issues and different processors making things difficult with little reward based on experience with another R package. However, I have used `fread` from the `data.table` package to decrease the runtime by ingesting the raw data more quickly. This resulted in a small, but notable decrease in the time necessary to run any of the functions.

Regarding the downloading issues, I've included two new functions that work with local files only and do not rely on R to download the files. Since CRAN guidelines specify that an R package should not write to disk (other than R's `tempfile()` or `tempdisk()`, I've elected to go this route so that a user may use an FTP client, web browser or other means to download and save the files locally before using this package in R. This also is useful for individuals who might desire to use this same data more than once, having the raw data on hand saving the download time and only importing into R from local.

Regarding the saving data when downloading, the functions now only download files that do not exist in the current R session. If you download temperature and create a stack, when creating a data frame of temperature will not trigger a new download unless you restart your R session.

Regarding the `with(CRU_data, plot(lon, lat))`, I've included examples using `ggplot2` to generate maps and graphs in README and vignette, this takes only a few seconds to generate on the machines I have a available to test on.

Regarding the 10 minute (second) issues. I've corrected the documentation and added the values both in arcminutes (the original from New et al.) and as a decimal value of degrees.

I've updated the paper.md file to be more detailed and addressed other comments regarding functionality and documentation as suggested.

Thank you both, @ldecicco-USGS and @ivanhanigan, again for the very constructive and helpful comments. They have helped me make this a much more robust R package that hopefully benefits R users.
