### 2017-02-26 9585573
Fix cran-comments/NEWS.md
