### 2017-02-27 a84c13e
Update README
Add documentation statement
