### 2017-02-28 #96 @sckott
Thanks for the changes @adamhsparks

@ivanhanigan @ldecicco-USGS are you two happy with the changes?
