### 2017-03-01 #96 @ivanhanigan
Thanks @adamhsparks for the changes.  I am not able to look at this for a
few days, and I cannot check on windows like @ldecicco-USGS reviewed with.
I'll get back to you as soon as I can, probably early next week.
Thanks again Adam.

On Wed, Mar 1, 2017 at 7:59 AM, Scott Chamberlain <notifications@github.com>
wrote:

> Thanks for the changes @adamhsparks <https://github.com/adamhsparks>
>
> @ivanhanigan <https://github.com/ivanhanigan> @ldecicco-USGS
> <https://github.com/ldecicco-USGS> are you two happy with the changes?
>
