### 2017-03-02 #96 @adamhsparks
Hello @ldecicco-USGS and @ivanhanigan, please don't rush to check these updates. I'm planning more.

I just found out about the `[rappdirs]`(https://github.com/hadley/rappdirs) package looking at @sckott's `ropensci::ccafs` package I'll modify the code to use a persistent cache directory for storing the files. This will take me a week or two I'd guess since I'm pretty busy with "real" work right now.

However, I think that this will provide a much nicer end-user experience.
