### 2017-03-02 #96 @sckott
thanks for the update @adamhsparks

see also https://github.com/ropensci/hoardr which is not quite to cran yet - wraps rappdirs to be a little more user friendly, see eg usage here https://github.com/ropenscilabs/cmipr/blob/master/R/onLoad.R and then you just export and document the caching object https://github.com/ropenscilabs/cmipr/blob/master/R/caching.R created in onload
