### 2017-03-08 #96 @ldecicco-USGS
I'll be happy to look again when @adamhsparks does his next round of updates (see 2 comments up). I think the first round only took ~1 hour.
