### 2017-03-13 92dc8ed
Break everything to fix it again
Set up using caching files downloaded in R using `rappdirs`

This is somewhat functional at this point. Cached files work, but
non-cached files will not. Need to determine how to check local files
vs requested and only return file list of requested files or download
what is necessary.
