+++ R/get_CRU_stack.R
@@ -26,29 +26,34 @@
 #' @details This function generates a raster stack object in R with the following
 #' possible fields as specified by the user:
 #' @param pre Logical. Fetch precipitation (millimetres/month) from server and
-#' return in a raster stack? Defaults to FALSE.
+#' return in the data frame? Defaults to \code{FALSE}.
 #' @param pre_cv Logical. Fetch cv of precipitation (percent) from server and
-#' return in a raster stack? Defaults to FALSE.
-#' @param rd0 Logical. Fetch wet-days (number days with >0.1 millimetres rain
-#' per month) and return in a raster stack? Defaults to FALSE.
+#' return in the data frame? Defaults to \code{FALSE}. NOTE. Setting this to
+#' \code{TRUE} will always results in \strong{pre} being set to \code{TRUE} and
+#' returned as well.
+#' @param rd0 Logical. Fetch wet-days (number days with >0.1millimetres rain per
+#' month) and return in the data frame? Defaults to \code{FALSE}.
 #' @param dtr Logical. Fetch mean diurnal temperature range (degrees Celsius)
-#' and return it in a raster stack? Defaults to FALSE.
+#' and return it in the data frame? Defaults to \code{FALSE}.
 #' @param tmp Logical. Fetch temperature (degrees Celsius) and return it in the
-#' raster stack? Defaults to FALSE.
+#' data frame? Defaults to \code{FALSE}.
 #' @param tmn Logical. Calculate minimum temperature values (degrees Celsius)
-#' and return it in a raster stack? Defaults to FALSE.
+#' and return it in the data frame? Defaults to \code{FALSE}.
 #' @param tmx Logical. Calculate maximum temperature (degrees Celsius) and
-#' return it in a raster stack? Defaults to FALSE.
-#' @param reh Logical. Fetch relative humidity and return it in a raster stack?
+#' return it in the data frame? Defaults to \code{FALSE}.
+#' @param reh Logical. Fetch relative humidity and return it in the data frame?
 #' Defaults to FALSE.
 #' @param sunp Logical. Fetch sunshine, percent of maximum possible (percent of
-#' day length) and return it in raster stack? Defaults to FALSE.
+#' day length) and return it in data frame? Defaults to \code{FALSE}.
 #' @param frs Logical. Fetch ground-frost records (number of days with ground-
-#' frost per month) and return it in raster stack? Defaults to FALSE.
+#' frost per month) and return it in data frame? Defaults to \code{FALSE}.
 #' @param wnd Logical. Fetch 10m wind speed (metres/second) and return it in the
-#' raster stack? Defaults to FALSE.
-#' @param elv Logical. Fetch elevation (converted to metres) and return it in a
-#' raster layer object? Defaults to FALSE.
+#' data frame? Defaults to \code{FALSE}.
+#' @param elv Logical. Fetch elevation (converted to metres) and return it in
+#' the data frame? Defaults to \code{FALSE}.
+#' @param cache Logical. Store CRU CL2.0 data files locally for later use? If
+#' \code{FALSE}, the downloaded files are removed when R session is closed.
+#' Defaults to \code{FALSE}.
 #'
 #' @examples
 #' # Download data and create a raster stack of precipitation and temperature.
@@ -58,6 +63,7 @@
 #' @seealso
 #' \code{\link{create_CRU_df}}
 #' \code{link{get_CRU_stack}}
+#' \code{\link{manage_CRU_cache}}
 #'
 #' @note
 #' This package automatically converts elevation values from kilometres to
@@ -82,8 +88,8 @@ get_CRU_stack <-
            sunp = FALSE,
            frs = FALSE,
[…]
