+++ NAMESPACE
@@ -1,5 +1,10 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,CRU_cache_info)
+export(CRU_cache_delete)
+export(CRU_cache_delete_all)
+export(CRU_cache_details)
+export(CRU_cache_list)
 export(create_CRU_df)
 export(create_CRU_stack)
 export(get_CRU_df)

[…]
