### 2017-03-14 c2dfad8
Add ability to cache data when downloaded in R
If user downloads data using R, data can be cached using `rappdirs`.

Additionally, when using data downloaded via FTP externally to R, the
files imported are now filtered and checked. Not all files in directory
will be imported, only those requested. Also, if a file is requested
but does not exist, R stops and returns error message.

Updates to documentation

Fixes to README to bring into line with CRAN policy regarding images,
use images from vignette so everything is self-contained
