+++ R/get_CRU_stack.R
@@ -23,46 +23,53 @@
 #'For more information see the description of the data provided by CRU,
 #'\url{https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt}
 #'
-#' @details This function generates a raster stack object in R with the following
-#' possible fields as specified by the user:
-#' @param pre Logical. Fetch precipitation (millimetres/month) from server and
-#' return in the data frame? Defaults to \code{FALSE}.
-#' @param pre_cv Logical. Fetch cv of precipitation (percent) from server and
-#' return in the data frame? Defaults to \code{FALSE}. NOTE. Setting this to
+#' @details This function generates a raster stack object in R with the
+#' following possible fields as specified by the user:
+#' @param pre Logical.  Fetch precipitation (millimetres/month) from server and
+#' return in the data frame?  Defaults to \code{FALSE}.
+#' @param pre_cv Logical.  Fetch cv of precipitation (percent) from server and
+#' return in the data frame?  Defaults to \code{FALSE}.  NOTE.  Setting this to
 #' \code{TRUE} will always results in \strong{pre} being set to \code{TRUE} and
 #' returned as well.
-#' @param rd0 Logical. Fetch wet-days (number days with >0.1millimetres rain per
-#' month) and return in the data frame? Defaults to \code{FALSE}.
-#' @param dtr Logical. Fetch mean diurnal temperature range (degrees Celsius)
-#' and return it in the data frame? Defaults to \code{FALSE}.
-#' @param tmp Logical. Fetch temperature (degrees Celsius) and return it in the
-#' data frame? Defaults to \code{FALSE}.
-#' @param tmn Logical. Calculate minimum temperature values (degrees Celsius)
-#' and return it in the data frame? Defaults to \code{FALSE}.
-#' @param tmx Logical. Calculate maximum temperature (degrees Celsius) and
-#' return it in the data frame? Defaults to \code{FALSE}.
-#' @param reh Logical. Fetch relative humidity and return it in the data frame?
+#' @param rd0 Logical.  Fetch wet-days (number days with >0.1millimetres rain
+#' per month) and return in the data frame? Defaults to \code{FALSE}.
+#' @param dtr Logical.  Fetch mean diurnal temperature range (degrees Celsius)
+#' and return it in the data frame?  Defaults to \code{FALSE}.
+#' @param tmp Logical.  Fetch temperature (degrees Celsius) and return it in the
+#' data frame?  Defaults to \code{FALSE}.
+#' @param tmn Logical.  Calculate minimum temperature values (degrees Celsius)
+#' and return it in the data frame?  Defaults to \code{FALSE}.
+#' @param tmx Logical.  Calculate maximum temperature (degrees Celsius) and
+#' return it in the data frame?  Defaults to \code{FALSE}.
+#' @param reh Logical.  Fetch relative humidity and return it in the data frame?
 #' Defaults to FALSE.
-#' @param sunp Logical. Fetch sunshine, percent of maximum possible (percent of
-#' day length) and return it in data frame? Defaults to \code{FALSE}.
+#' @param sunp Logical.  Fetch sunshine, percent of maximum possible (percent of
+#' day length) and return it in data frame?  Defaults to \code{FALSE}.
 #' @param frs Logical. Fetch ground-frost records (number of days with ground-
-#' frost per month) and return it in data frame? Defaults to \code{FALSE}.
-#' @param wnd Logical. Fetch 10m wind speed (metres/second) and return it in the
-#' data frame? Defaults to \code{FALSE}.
[…]
