+++ R/create_CRU_df.R
@@ -73,29 +73,95 @@
 #' metres.
 #'
 #' @export
-create_CRU_df <- function(pre = FALSE,
-                          pre_cv = FALSE,
-                          rd0 = FALSE,
-                          tmp = FALSE,
-                          dtr = FALSE,
-                          reh = FALSE,
-                          tmn = FALSE,
-                          tmx = FALSE,
-                          sunp = FALSE,
-                          frs = FALSE,
-                          wnd = FALSE,
-                          elv = FALSE,
-                          dsn = "") {
+create_CRU_df <-   function(pre = FALSE,
+                            pre_cv = FALSE,
+                            rd0 = FALSE,
+                            tmp = FALSE,
+                            dtr = FALSE,
+                            reh = FALSE,
+                            tmn = FALSE,
+                            tmx = FALSE,
+                            sunp = FALSE,
+                            frs = FALSE,
+                            wnd = FALSE,
+                            elv = FALSE,
+                            dsn = "") {
+
   if (!isTRUE(pre) & !isTRUE(pre_cv) & !isTRUE(rd0) & !isTRUE(tmp) &
       !isTRUE(dtr) & !isTRUE(reh) & !isTRUE(tmn) & !isTRUE(tmx) &
       !isTRUE(sunp) & !isTRUE(frs) & !isTRUE(wnd) & !isTRUE(elv)) {
-    stop("\nYou must select at least one element for import.\n")
+    stop("\nYou must select at least one element for importing.\n")
   }
 
   .validate_dsn(dsn)
 
+  # check if pre_cv or tmx/tmn (derived) are true, make sure proper ----------
+  # parameters set TRUE
+  if (isTRUE(pre_cv)) {
+    pre <- TRUE
+  }
+
+  if (isTRUE(tmn) | isTRUE(tmx)) {
+    dtr <- tmp <- TRUE
+  }
+
+  dtr_file <- "grid_10min_dtr.dat.gz"
+  tmp_file <- "grid_10min_tmp.dat.gz"
+  reh_file <- "grid_10min_reh.dat.gz"
+  elv_file <- "grid_10min_elv.dat.gz"
+  pre_file <- "grid_10min_pre.dat.gz"
+  sun_file <- "grid_10min_sunp.dat.gz"
+  wnd_file <- "grid_10min_wnd.dat.gz"
+  frs_file <- "grid_10min_frs.dat.gz"
+  rd0_file <- "grid_10min_rd0.dat.gz"
+
+  object_list <- c(dtr, tmp, reh, elv, pre, sunp, wnd, frs, rd0)
+
   files <-
-    list.files(dsn, pattern = ".dat.gz$", full.names = TRUE)
+    c(
+      dtr_file,
+      tmp_file,
+      reh_file,
+      elv_file,
+      pre_file,
+      sun_file,
+      wnd_file,
+      frs_file,
+      rd0_file
+    )
+  names(files) <-
+    names(object_list) <-
+    c(
+      "dtr_file",
+      "tmp_file",
+      "reh_file",
+      "elv_file",
+      "pre_file",
+      "sun_file",
+      "wnd_file",
+      "frs_file",
+      "rd0_file"
+    )
+
+  # filter files -------------------------------------------------------------
+  # which files are being requested?
+  files <- files[object_list %in% !isTRUE(files)]
+
+  # filter files from cache directory in case there are local files for which
+  # we do not want data
+  dsn_contents <- as.list(list.files(dsn, pattern = ".dat.gz$"))
+
+  files <- dsn_contents[dsn_contents %in% files]
+
+  if (length(files) < 0) {
[…]
