### 2017-03-14 204c029
Consolidate code from create_* functions
