+++ R/create_CRU_stack.R
@@ -100,76 +100,19 @@ create_CRU_stack <- function(pre = FALSE,
 
   .validate_dsn(dsn)
 
-  # check if pre_cv or tmx/tmn (derived) are true, make sure proper ----------
-  # parameters set TRUE
-  if (isTRUE(pre_cv)) {
-    pre <- TRUE
-  }
-
-  if (isTRUE(tmn) | isTRUE(tmx)) {
-    dtr <- tmp <- TRUE
-  }
-
-  dtr_file <- "grid_10min_dtr.dat.gz"
-  tmp_file <- "grid_10min_tmp.dat.gz"
-  reh_file <- "grid_10min_reh.dat.gz"
-  elv_file <- "grid_10min_elv.dat.gz"
-  pre_file <- "grid_10min_pre.dat.gz"
-  sun_file <- "grid_10min_sunp.dat.gz"
-  wnd_file <- "grid_10min_wnd.dat.gz"
-  frs_file <- "grid_10min_frs.dat.gz"
-  rd0_file <- "grid_10min_rd0.dat.gz"
-
-  object_list <- c(dtr, tmp, reh, elv, pre, sunp, wnd, frs, rd0)
-
-  files <-
-    c(
-      dtr_file,
-      tmp_file,
-      reh_file,
-      elv_file,
-      pre_file,
-      sun_file,
-      wnd_file,
-      frs_file,
-      rd0_file
-    )
-  names(files) <-
-    names(object_list) <-
-    c(
-      "dtr_file",
-      "tmp_file",
-      "reh_file",
-      "elv_file",
-      "pre_file",
-      "sun_file",
-      "wnd_file",
-      "frs_file",
-      "rd0_file"
-    )
-
-  # filter files -------------------------------------------------------------
-  # which files are being requested?
-  files <- files[object_list %in% !isTRUE(files)]
-
-  # filter files from cache directory in case there are local files for which
-  # we do not want data
-  dsn_contents <- as.list(list.files(dsn, pattern = ".dat.gz$"))
-
-  files <- dsn_contents[dsn_contents %in% files]
-
-  if (length(files) < 0) {
-    stop(
-      "You are requesting a file that you have not downloaded or is not in the directory specified."
-    )
-  }
-
-  # add full file path to the files
-  files <- paste0(dsn, "/", files)
-
-  # fill the space with a "\" for R, if one exists
-  files <- gsub(" ", "\\ ", files, fixed = TRUE)
-
+  files <- .get_local(pre,
+                      pre_cv,
+                      rd0,
+                      tmp,
+                      dtr,
+                      reh,
+                      tmn,
+                      tmx,
+                      sunp,
+                      frs,
+                      wnd,
+                      elv,
+                      cache_dir = dsn)
   s <- create_stacks(tmn, tmx, tmp, dtr, pre, pre_cv, files)
   return(s)
 }
