### 2017-03-14 6132152
minor corrections on cache management
