### 2017-03-14 c2a53b2
More minor corrections in cache management
