### 2017-03-14 173f759
Better tests for create_CRU_df()
