### 2017-03-14 e48971c
Improve tests for `create_CRU_stack()`
