+++ tests/testthat/test-create_CRU_stack.R
@@ -1,3 +1,5 @@
+
+
 context("create_CRU_stack")
 
 test_that("create_CRU_stack fails if no dsn is specified", {
@@ -465,30 +467,71 @@ test_that("create_CRU_df lists only .dat.gz files in the given dsn", {
       7.6
     )
   )
-  gz1 <- gzfile(paste0(tempdir(), "/grid_10min_pre.gz"), "w")
-  utils::write.table(
-    pre_data,
-    file = gz1,
-    col.names = FALSE,
-    row.names = FALSE
-  )
+  gz1 <- gzfile(paste0(tempdir(), "/grid_10min_pre.dat.gz"), "w")
+  utils::write.table(pre_data,
+                     file = gz1,
+                     col.names = FALSE,
+                     row.names = FALSE)
   close(gz1)
 
   gz1 <- gzfile(paste0(tempdir(), "/grid_10min_tmp.dat.gz"), "w")
-  utils::write.table(
-    tmp_data,
-    file = gz1,
-    col.names = FALSE,
-    row.names = FALSE
-  )
+  utils::write.table(tmp_data,
+                     file = gz1,
+                     col.names = FALSE,
+                     row.names = FALSE)
   close(gz1)
 
-  dsn <- tempdir()
+  pre <- TRUE
+  pre_cv <- TRUE
+  rd0 <- FALSE
+  tmp <- TRUE
+  dtr <- FALSE
+  reh <- FALSE
+  tmn <- FALSE
+  tmx <- FALSE
+  sunp <- FALSE
+  frs <- FALSE
+  wnd <- FALSE
+  elv <- FALSE
+  cache_dir <- tempdir()
 
-  files <-
-    list.files(dsn, pattern = ".dat.gz$", full.names = TRUE)
+  files <- .get_local(pre,
+                      pre_cv,
+                      rd0,
+                      tmp,
+                      dtr,
+                      reh,
+                      tmn,
+                      tmx,
+                      sunp,
+                      frs,
+                      wnd,
+                      elv,
+                      cache_dir)
 
-  expect_type(files, "character")
-  expect_equal(files, paste0(tempdir(), "/grid_10min_tmp.dat.gz"))
+  expect_equal(length(files), 2)
+  expect_equal(basename(files[1]), "grid_10min_pre.dat.gz")
+  expect_equal(basename(files[2]), "grid_10min_tmp.dat.gz")
+
+  s <- create_stacks(tmn, tmx, tmp, dtr, pre, pre_cv, files)
+
+  expect_named(s, c("pre", "tmp"))
+  expect_equal(
+    raster::cellStats(s$tmp, max),
+    c(
+      8.6,
+      8.6,
+      7.3,
+      5.6,
+      3.4,
+      2.0,
+      1.5,
+      2.2,
+      3.6,
+      5.4,
+      6.7,
+      8.0
+    )
+  )
+  expect_type(s, "list")
 })
-
