### 2017-03-15 e0f2963
Improved tests for `get_CRU_*()` functions
