### 2017-03-15 c0620c7
New test for `.validate_dsn()`
