### 2017-03-15 2fb1337
Clean up tests
Mainly formatting and removing duplicates and cleaning up tests to be
more readable and easier to follow
