+++ tests/testthat/test-get_CRU_df.R
@@ -1,10 +1,14 @@
 
 context("get_CRU_df")
 
+# Test that get_CRU_df fails if no parameters are TRUE -------------------------
+
 test_that("get_CRU_df fails if no parameters are TRUE", {
   expect_error(get_CRU_df(),
                "You must select at least one element for download.")
 })
+
+# Test that create_CRU_df lists only .dat.gz files in the given dsn ------------
 
 test_that("create_CRU_df lists only .dat.gz files in the given dsn", {
   # create files for testing, these data are the first 10 lines of pre and tmp
@@ -488,6 +492,8 @@ test_that("create_CRU_df lists only .dat.gz files in the given dsn", {
   expect_equal(files, paste0(tempdir(), "/grid_10min_tmp.dat.gz"))
 })
 
+# Test that get_CRU_df sets the cache directory properly when cache is TRUE ----
+
 test_that("get_CRU_df sets the cache directory properly when cache is TRUE",
           {
             cache <- TRUE
@@ -504,6 +510,8 @@ test_that("get_CRU_df sets the cache directory properly when cache is TRUE",
             expect_equal(cache_dir, rappdirs::user_config_dir("getCRUCLdata"))
 
           })
+
+# Test that get_CRU_df sets the cache dir properly when cache is FALSE ---------
 
 test_that("get_CRU_df sets the cache directory properly when cache is FALSE",
           {
