### 2017-03-15 907125c
Better tests, more readable now
Minor change in `manage_CRU_cache` functionality where deleting a single file only requires the filename not the whole filepath.

Update documentation as well
