+++ tests/testthat/test-validate_dsn.R
@@ -3,6 +3,7 @@ context(".validate_dsn")
 # Test that .validate_dsn stops if the dsn is not provided ---------------------
 
 test_that(".validate_dsn stops if the dsn is not provided", {
+  skip_on_cran()
   dsn <- NULL
   expect_error(.validate_dsn(dsn),
          "You must define the dsn where you have stored the local files
