+++ tests/testthat/test-get_CRU_stack.R
@@ -493,6 +493,7 @@ test_that("create_CRU_df lists only .dat.gz files in the given dsn", {
 
 test_that("get_CRU_stack sets the cache directory properly when cache is TRUE",
           {
+            skip_on_cran()
             cache <- TRUE
 
             if (isTRUE(cache)) {
@@ -511,6 +512,7 @@ test_that("get_CRU_stack sets the cache directory properly when cache is TRUE",
 
 test_that("get_CRU_stack sets the cache directory properly when cache is FALSE",
           {
+            skip_on_cran()
             cache <- FALSE
 
             if (isTRUE(cache)) {
