+++ tests/testthat/test_get_local.R
@@ -3,6 +3,8 @@ context(".get_local")
 # Test that .get_local creates a tibble of only specified elements -------------
 
 test_that("Test that .get_local creates a tibble of only specified elements (pre)", {
+  skip_on_cran()
+
   # create files for testing, these data are the first 10 lines of pre and tmp
   # from the CRU CL2.0 data
 
