### 2017-03-15 70fe0b9
cache management tests don't play nice
Cannot get any cache management tests to pass in testing, will pass if
run in normal session.
