+++ tests/testthat/test_CRU_cache_management.R
@@ -1,485 +1,0 @@
-
-context("manage_CRU_cache")
-
-# Test that manage_CRU_cache functions work as expected -------------------------
-test_that("Test that manage_CRU_cache functions work as expected",
-          {
-            unlink(
-              list.files(
-                path = rappdirs::user_config_dir("getCRUCLdata"),
-                pattern = ".dat.gz$",
-                full.names = TRUE
-              )
-            )
-
-            pre_data <- rbind(
-              c(
-                -59.083,
-                -26.583,
-                105.4,
-                121.3,
-                141.3,
-                146.7,
-                159.6,
-                162.4,
-                141.5,
-                151.1,
-                141.6,
-                124.9,
-                110.0,
-                93.9,
-                35.2,
-                38.7,
-                38.4,
-                27.5,
-                49.5,
-                40.8,
-                50.8,
-                33.5,
-                42.2,
-                56.6,
-                35.5,
-                43.4
-              ),
-              c(
-                -58.417,
-                -26.250,
-                106.3,
-                121.3,
-                141.6,
-                147.4,
-                160.3,
-                163.5,
-                142.1,
-                151.7,
-                142.1,
-                125.5,
-                110.5,
-                94.4,
-                36.2,
-                39.8,
-                38.6,
-                27.8,
-                49.6,
-                40.8,
-                51.0,
-                33.3,
-                43.1,
-                57.2,
-                35.8,
-                44.2
-              ),
-              c(
-                -58.417,
-                -26.417,
-                106.6,
-                121.9,
-                142.1,
-                147.5,
-                160.3,
-                163.4,
-                142.0,
-                151.7,
-                142.1,
-                125.5,
-                110.5,
-                94.5,
-                36.2,
-                39.7,
-                38.6,
-                27.8,
-                49.7,
-                40.9,
-                51.0,
-                33.3,
-                43.1,
-                57.2,
-                35.9,
-                44.2
-              ),
-              c(
-                -55.917,
-                -67.250,
-                73.1,
-                78.7,
-                81.5,
-                81.2,
-                68.5,
-                67.9,
-                62.5,
-                60.2,
-                60.4,
-                55.0,
-                59.9,
-                71.2,
-                44.1,
-                54.2,
-                51.1,
-                54.2,
-                59.0,
-                47.2,
-                62.0,
-                65.6,
-                58.9,
-                47.8,
-                50.9,
-                47.0
-              ),
-              c(
-                -55.750,
-                -67.250,
-                68.7,
-                73.8,
-                76.6,
-                76.9,
-                65.3,
-                64.4,
-                58.6,
-                56.9,
-                56.7,
-                51.5,
-                56.4,
-                67.1,
-                44.7,
-                55.0,
-                51.9,
-                54.4,
-                59.6,
-                48.2,
-                63.0,
-                66.5,
-                60.3,
-                48.8,
-                51.7,
-                47.6
-              ),
-              c(
-                -55.750,
-                -67.417,
-                70.2,
-                75.7,
-                78.8,
-                78.6,
-                66.5,
-                66.2,
-                59.8,
-                58.7,
-                58.1,
-                52.7,
-                57.2,
-                68.4,
-                44.6,
-                55.0,
-                51.9,
-                54.8,
-                59.8,
-                48.2,
-                62.9,
-                66.4,
-                60.1,
-                48.5,
-                51.9,
-                47.6
-              ),
-              c(
-                -55.750,
-                -67.583,
-                71.7,
-                77.4,
-                80.6,
-                81.0,
-                68.6,
-                67.8,
-                61.8,
-                60.3,
-                59.8,
[…]
