### 2017-03-15 #96 @adamhsparks
Ok folks. I've had another go at this.

I've added the ability to cache files in the users' home filespace as I mentioned and cleaned up a few other minor items while I was at it.

Please have another look and let me know of any issues or suggestions.
