### 2017-03-15 d0bba72
Update reference to CRU data
The data were inconsistently referred to throughout the package. I've
tried to refer to the data as "CRU CL v. 2.0" throughout the package
and documentation now.
