+++ README.Rmd
@@ -27,7 +27,7 @@ knitr::opts_chunk$set(
 Author/Maintainer: Adam Sparks
 
 The getCRUCLdata package provides four functions that automate importing CRU 
-CL2.0 climatology data into R, facilitate the calculation of minimum temperature
+CL v. 2.0 climatology data into R, facilitate the calculation of minimum temperature
 and maximum temperature, and formats the data into a 
 [tidy data frame](http://vita.had.co.nz/papers/tidy-data.html) or a list
 of [raster stack](https://cran.r-project.org/web/packages/raster/vignettes/Raster.pdf) objects for use in an R session. CRU CL v. 2.0 data are a gridded
