### 2017-03-15 #96 @ivanhanigan
Adam.
The new README looks great. I'll try to get to this in the next few days.
