### 2017-03-26 #96 @ivanhanigan
Hi Adam thanks for the new release of the `getCRUCLdata` package.
I started a new review BUT have not done a comprehensive review yet.

But thought I'd just check in on something I think is worth discussing re cached files of a one-off download of the data vs repeated downloading on-the-fly.

When I looked at the code in the readme, it was obvious `~/Downloads' does not work.
```{r}
t <- create_CRU_df(tmp = TRUE, dsn = "~/Downloads")
```

So I checked the function and found downloading is actually to this directory:
```{r}
dsn <- "/home/ivan_hanigan/.config/getCRUCLdata"
# which works juset as well
t2 <- create_CRU_df(tmp = TRUE, dsn = dsn)
identical(t, t2)
# TRUE
```

It seems to me that this location will not be very clear to users of the program if asked `where is the CRU data you downloaded as a one-off'?  It seems like the package assumes people will download the data whenever the function is invoked - 'on-the-fly'.

My concern I wanted to raise before proceeding with my review is that data downloading as a 'one-off' might considered more important by the user, and being prompted to choose a place to store the result for further management, versus a 'on-the-fly' data download to a temporary/system location crated for the current session.
This assumes good network connections and there are associated issues with online data provision (including trust that the data provider will not change the data at the online location between downloads) that require soem consideration .

So I just thought I'd flag this issue now and suggest the following:

What about making ` get_CRU` to ask the User where to download the CRU data to?
Then this makes the location explicit and makes the dataset a more recognisable contribution to the files that they have on their system.

PS I noticed a duplicated chunk of text in this
```{r}
?get_CRU_df
"Duplicated?
# Download data and create a data frame of precipitation and temperature
# without caching the data files
CRU_pre_tmp <- get_CRU_df(pre = TRUE, tmp = TRUE)

# Download data and create a data frame of precipitation and temperature
# without caching the data files
CRU_pre_tmp <- get_CRU_df(pre = TRUE, tmp = TRUE)
"
```
All the best, and I will continue the review this week - in between other work responsibilities.

Thanks Adam!

I
