### 2017-03-26 7dff166
Clarify the difference between functions some more in README
Add more to description of `create_CRU_df()` in README
