### 2017-03-26 #96 @adamhsparks
Hi @ivanhanigan,
I don't understand why using `create_CRU_df(dsn = "~/Downloads")` won't work as I've illustrated?

It seems to work properly for me, could you give me the line numbers in the code that you think are incorrect? `create_CRU_df()` does not download anything. It requires the data to already exist locally, e.g "~/Downloads".  `get_CRU_df()` downloads to the location you've described and caches.

Is there a need to rename the functions so that the differences are more clear?

PS
You're right. It was a duplicate. I've fixed that.
