### 2017-03-27 #96 @ivanhanigan
Hi,
It is not incorrect but it depends on the user moving the data from the cache_dir to the folder `~/Downloads` in your example.

```
#in README it says
t <- create_CRU_df(tmp = TRUE, dsn = "~/Downloads")
# but the function get_CRU_df() downloads and caches data in a special directory
# so you probably need to explain to the user
# they need to move/copy the data to the folder they then specified to the dsn argument
# we can find this with
cachedir <- rappdirs::user_config_dir("getCRUCLdata")
# on my system this is "/home/ivan_hanigan/.config/getCRUCLdata"
# so I can move the files from there to the correct folder, or leave them there
# (but that seems like sloppy data management to me)
# I just move them to a data folder here
system(sprintf("mv %s/ data_provided/", cachedir))
# and then when I come back to use them I know they are handy
t3 <- create_CRU_df(tmp = TRUE, dsn = "data_provided")
# and identical to what I downloaded fresh with
t <- get_CRU_df(tmp = TRUE, cache = FALSE)
identical(t, t3)
# TRUE
```
Or do you recommend that users store the raw data in the `.config/getCRUCLdata` location?
