### 2017-03-27 #96 @adamhsparks
@ivanhanigan, it's not intended for the user to move data at all. It's intended if a user has network issues, to use an external FTP program was previously suggested.

The caching works only with the `get_CRU_df()` function. If you use that function, you would need to re-use it to take advantage of the caching functionality. Perhaps I should just drop the caching since it seems to be creating confusion?
