### 2017-03-27 3048237
Attempt to clarify differences between `get_CRU*` and `create_CRU*`
Add more description in README based on ivanhanigan’s feedback in
rOpenSci review to try to clarify when and how the functions are used
