### 2017-03-27 0897c78
Spelling corrections in documentation
