+++ R/get_CRU_df.R
@@ -54,7 +54,7 @@
 #' @param cache Logical.  Store CRU CL v. 2.0 data files locally for later use?
 #' If \code{FALSE}, the downloaded files are removed when R session is closed.
 #' To take advantage of cached files in future sessions, use \code{cache = TRUE}
-#' after the inital download and caching.  Defaults to \code{FALSE}.
+#' after the initial download and caching.  Defaults to \code{FALSE}.
 #'
 #' @examples
 #' \dontrun{
