### 2017-03-27 #96 @adamhsparks
I've tried to revise the README text for clarity.
