### 2017-03-27 #96 @ldecicco-USGS
I pulled the latest changes from the master branch here and built from that. When I tried the first example in the vignette:

```
CRU_data <- create_CRU_df(pre = TRUE,
                          pre_cv = TRUE,
                          rd0 = TRUE,
                          tmp = TRUE,
                          dtr = TRUE,
                          reh = TRUE,
                          tmn = TRUE,
                          tmx = TRUE,
                          sunp = TRUE,
                          frs = TRUE,
                          wnd = TRUE,
                          elv = TRUE)
 Error in .validate_dsn(dsn) :
File directory does not exist: .
```

I thought maybe that was because the vignette maybe is dated? So I tried the first example for that function:
```
CRU_pre_tmp <- create_CRU_df(pre = TRUE, tmp = TRUE, dsn = "~/Downloads")
Error in .validate_dsn(dsn) :
File directory does not exist: ~/Downloads.
```
So...I need to create some actual path in the dsn argument. I did this:
```
CRU_pre_tmp <- create_CRU_df(pre = TRUE, tmp = TRUE, dsn = "D:/LADData")

Creating data frame now.

  |                                                     |   0%gzip: D:/LADData/ is a directory -- ignored
 Show Traceback

 Rerun with Debug
 Error in data.table::fread(paste0("gzip -dc ", .files), header = FALSE) :
  File is empty: C:\Users\ldecicco\AppData\Local\Temp\1\RtmpszBo52\filefac5fb84c54 In addition: Warning messages:
1: running command 'C:\Windows\system32\cmd.exe /c (gzip -dc D:/LADData/) > C:\Users\ldecicco\AppData\Local\Temp\1\RtmpszBo52\filefac5fb84c54' had status 2
2: In shell(paste("(", input, ") > ", tt, sep = "")) :
  '(gzip -dc D:/LADData/) > C:\Users\ldecicco\AppData\Local\Temp\1\RtmpszBo52\filefac5fb84c54' execution failed with error code 2
```

So, I think I agree with @ivanhanigan that the documentation on what I'm suppose to put in the `dsn` argument needs to be improved.
