### 2017-03-28 #96 @adamhsparks
I don't know what's happened here. That's not the first example in the vignette and the function arguments are not all supplied properly so it won't run and stops as it should?

This is the first example in the vignette:

```r
library(getCRUCLdata)

CRU_data <- get_CRU_df(pre = TRUE,
                       pre_cv = TRUE,
                       rd0 = TRUE,
                       tmp = TRUE,
                       dtr = TRUE,
                       reh = TRUE,
                       tmn = TRUE,
                       tmx = TRUE,
                       sunp = TRUE,
                       frs = TRUE,
                       wnd = TRUE,
                       elv = TRUE,
                       cache = TRUE)
```
