### 2017-03-28 df9ffad
Attempt to enhance documentation further
Trying to better describe how to use the cache functionality of the
`get_CRU_*()` functions.
