### 2017-03-28 ef058fc
More documentation changes, mainly in function files
