### 2017-03-28 #96 @adamhsparks
@ldecicco-USGS,
I've added more description to the documentation, but I still don't understand the first example that you posted. I don't know where that comes from?

Thanks!
