### 2017-03-28 #96 @ivanhanigan
Hi again,
I have re-read the README section "Creating tidy data frames for use in R"
and it does make a lot of sense thanks.
I appreciate the way the package is trying to accommodate a variety of
different data downloading/storage/networking scenarios.
I think that you've done a great job of making this flexible for that, but
maybe the additional detail that needs to be discussed for
caching/FTP-FileZilla is confusing and could be put in a later section such
as "advanced usage/special circumstances" section and the more visible
sections of README and vignettes just show the simple application of
running the `get_CRU_df()` function.

Then you could move into a discussion about how useless repeatedly
downloading copies of the same data, OR internet connectivity problems
might lead you to want to use FileZilla or caching?

I'll move on to the next components of the documentation tomorrow.  I think
that if you don't want to make more changes then the current approach to
caching/downloading is adequate.  Up to you mate!

Cheers, Ivan

On Tue, Mar 28, 2017 at 1:13 PM, Adam H. Sparks <notifications@github.com>
wrote:

> @ldecicco-USGS <https://github.com/ldecicco-USGS>,
> I've added more description to the documentation, but I still don't
> understand the first example that you posted. I don't know where that comes
> from?
>
> Thanks!
>
