### 2017-03-28 #96 @ldecicco-USGS
OK, I think what was happening with my first example is that I was using the older version of the vignette. I re-built the vignettes and got things working.

I think where there could be confusion (there was from me at least...), that might easily be straightened out for the user, is have the very first example in the vignette be how to actually download the data. The vignette starts with "Using getCRUCLdata", which is definitely the cool part of the package, but maybe having a section above that "Getting CRUCL Data" with the explicit instruction downloading for the first time would be helpful. Then, all the examples with `cache=TRUE` immediately work.

Otherwise, looks good!

A minor suggestion that might help people like me (who..maybe...have been known to only skim through the instructions :sweat: )....I kept trying to run `create_CRU_df` without having run `get_CRU_df` first. I saw the "dsn" argument and thought it would _save_ the data there, not _find_ the data there. Perhaps you could put a custom error message in that function, that if the `dsn` argument doesn't have the correct files, to use `get_CRU_df` to download the data?
