### 2017-03-28 #96 @adamhsparks
@ldecicco-USGS that's a good suggestion, I'm guilty of skimming documentation too. I'll implement that change.

I'll reorganise the documentation as you and Ivan have suggested. You've both been very helpful, thank you.
