### 2017-03-30 783115f
Updates to documentation
Split more advanced functions out into “Advanced Usage” and improve
header level usage
