+++ R/get_CRU.R
@@ -98,7 +98,8 @@
         },
         error = function(x) {
           do.call(file.remove, list(list.files(cache_dir, full.names = TRUE)))
-          stop("\nThe file downloads have failed. Please start the download again.\n")
+          stop("\nThe file downloads have failed.\n
+               \nPlease start the download again.\n")
         }
       )
     }
@@ -205,4 +206,3 @@
 
   return(files)
 }
-
