+++ tests/testthat/test-get_CRU_df.R
@@ -494,8 +494,7 @@ test_that("create_CRU_df lists only .dat.gz files in the given dsn", {
 
 # Test that get_CRU_df sets the cache directory properly when cache is TRUE ----
 
-test_that("get_CRU_df sets the cache directory properly when cache is TRUE",
-          {
+test_that("get_CRU_df sets the cache directory properly when cache is TRUE", {
             skip_on_cran()
 
             cache <- TRUE
@@ -515,8 +514,7 @@ test_that("get_CRU_df sets the cache directory properly when cache is TRUE",
 
 # Test that get_CRU_df sets the cache dir properly when cache is FALSE ---------
 
-test_that("get_CRU_df sets the cache directory properly when cache is FALSE",
-          {
+test_that("get_CRU_df sets the cache directory properly when cache is FALSE", {
 
             skip_on_cran()
             cache <- FALSE
