### 2017-03-30 #96 @adamhsparks
Hi folks, I've edited the README and vignette to include a quick start and a more advanced section on caching and using `create_CRU_*()` functions.

No changes to functionality, only documentation.

I've also linted the package and fixed some errors in the documentation with typos, "a a", etc.
