### 2017-03-31 512058e
Update NEWS.md and cran-comments.md
