+++ cran-comments.md
@@ -11,16 +11,17 @@
 
 ## Major Changes
 
-  * `create_CRU_stack()` and `create_CRU_df()` now only work with locally available files. If you need to fetch and create a data frame or raster stack of the data, please use the new functions, `get_CRU_stack()` and `get_CRU_stack()`
-  * R >=3.2.0 now required
+  * `create_CRU_stack()` and `create_CRU_df()` now only work with locally available files. If you need to fetch and create a data frame or raster stack of the data, please use the new functions, `get_CRU_stack()` and `get_CRU_stack()`  
+  * R >=3.2.0 now required  
+  * Data can be cached using either `get_CRU_stack()` or `get_CRU_df()` for later use  
 
 ## Minor Changes
 
   * Improved documentation with examples on mapping and graphing and more detail regarding the data itself
-  * Change the method in which files are downloaded to use `httr::GET()`
-  * Ingest data using `data.table::fread` to decrease the amount of time necessary to run the functions
-  * Functions check to see if data file(s) have already been downloaded during current R session, if so data file(s) are not requested for download again
-  * Months are returned as a factor object in the tidy data frame
+  * Change the method in which files are downloaded to use `httr::GET()`  
+  * Ingest data using `data.table::fread` to decrease the amount of time necessary to run the functions  
+  * Functions check to see if data file(s) have already been downloaded during current R session, if so data file(s) are not requested for download again  
+  * Months are returned as a factor object in the tidy data frame  
   
 ## Reverse dependencies
 
