### 2017-03-31 #96 @sckott
is all taken care of @adamhsparks ? If so, i'll take a final look
