### 2017-03-31 #96 @adamhsparks
@sckott, I hope that I've addressed @ldecicco-USGS' and @ivanhanigan's comments re: documentation.

I'm still trying to change the error message when the DSN doesn't exist or files aren't found. The error stops and works, but the issue I'm having is with getting it to pass the test. :/
