### 2017-04-01 6add101
Fix citation; update title; fix formatting of getCRUCLdata
