+++ R/validate_dsn.R
@@ -2,7 +2,8 @@
 .validate_dsn <- function(dsn) {
   if (is.null(dsn)) {
     stop("You must define the dsn where you have stored the local files
-         for import.")
+         for import. If you want to download files using R, use one of the
+         'get_CRU_*()' functions provided.")
   } else {
     dsn <- trimws(dsn)
     if (substr(dsn, nchar(dsn) - 1, nchar(dsn)) == "//") {
