### 2017-04-01 a1bb181
fix test for .validate_dsn()
