### 2017-04-01 #96 @adamhsparks
@sckott, I think we're good. I've tidied it up a bit more here and there. Have a look now when you have time.
