### 2017-04-02 3613e6a
Add `#`@return` to help files
