### 2017-04-02 61648b7
Documentation corrections
