+++ R/create_CRU_df.R
@@ -28,8 +28,6 @@
 #'For more information see the description of the data provided by CRU,
 #'\url{https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt}
 #'
-#' @details This function generates a data.frame object in R with the following
-#' possible fields as specified by the user:
 #' @param pre Logical. Fetch precipitation (millimetres/month) from server and
 #'  return in the data frame? Defaults to FALSE.
 #' @param pre_cv Logical. Fetch cv of precipitation (percent) from server and
@@ -66,7 +64,6 @@
 #'
 #' @seealso
 #' \code{\link{get_CRU_df}}
-#' \code{\link{get_CRU_stack}}
 #'
 #' @return A tidy data frame of CRU CL v. 2.0 climatology elements as a
 #' \code{\link[tibble]{tibble}} object
