+++ man/create_CRU_df.Rd
@@ -81,10 +81,6 @@ Nomenclature and units from readme.txt:
 For more information see the description of the data provided by CRU,
 \url{https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt}
 }
-\details{
-This function generates a data.frame object in R with the following
-possible fields as specified by the user:
-}
 \note{
 This package automatically converts elevation values from kilometres to
 metres.
@@ -99,5 +95,4 @@ CRU_pre_tmp <- create_CRU_df(pre = TRUE, tmp = TRUE, dsn = "~/Downloads")
 }
 \seealso{
 \code{\link{get_CRU_df}}
-\code{\link{get_CRU_stack}}
 }
