+++ man/get_CRU_stack.Rd
@@ -82,10 +82,6 @@ Nomenclature and units from readme.txt:
 For more information see the description of the data provided by CRU,
 \url{https://crudata.uea.ac.uk/cru/data/hrg/tmc/readme.txt}
 }
-\details{
-This function generates a raster stack object in R with the
-following possible fields as specified by the user:
-}
 \note{
 This package automatically converts elevation values from kilometres to
 metres.
@@ -108,7 +104,6 @@ CRU_tmp <- get_CRU_stack(tmp = TRUE, tmn = TRUE, tmx = TRUE, cache = TRUE)
 }
 }
 \seealso{
-\code{\link{create_CRU_df}}
-\code{\link{get_CRU_stack}}
+\code{\link{create_CRU_stack}}
 \code{\link{manage_CRU_cache}}
 }
