### 2017-04-02 2db964a
Add author to references
