### 2017-04-02 cc38208
Fix documentation for manage_CRU_cache()
