### 2017-04-03 4876755
Move README figures to man/figures per CRAN
