### 2017-04-04 #96 @adamhsparks
I do. I've added the footer. I'll wait for the transfer request and update all the links and submit this latest version to CRAN.

Thanks for your time, suggestion and effort @ivanhanigan, @ldecicco-USGS  and @sckott. Once again, the rOpenSci review process really made the package a much better offering.
