### 2017-04-04 1786d5e
Update documentation
Fix typos, formatting issues, unclear wording, etc.
