+++ R/manage_cached_files.R
@@ -3,7 +3,7 @@
 #' @description The user is given an option when downloading the CRU CL v. 2.0
 #' data to cache or not to cache the data for later use.  If
 #' \code{cache == TRUE}, then the CRU CL v. 2.0 data files are saved in a
-#' directory in the users' home files pace.  These functions provide facilities
+#' directory in the users' home file space.  These functions provide facilities
 #' for interacting and managing these files.
 #'
 #' @export
