### 2017-04-04 fe12425
Correct raster link with proper CRAN approved link
