### 2017-04-04 d25539c
Fix title case in DESCRIPTION; update cran-comments.md
