+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: getCRUCLdata
 Type: Package
-Title: Use and explore CRU CL v. 2.0 Climatology Elements in R
+Title: Use and Explore CRU CL v. 2.0 Climatology Elements in R
 Version: 0.1.5
 Authors@R: person(given = "Adam", family = "Sparks", 
                   email = "adamhsparks@gmail.com", role = c("aut", "cre"))
