### 2017-04-04 3bb2dd9
Replace adamhsparks with ropensci to move repository over to ropensci
