+++ README.Rmd
@@ -12,9 +12,9 @@ knitr::opts_chunk$set(
 )
 ```
 
-[![Travis-CI Build Status](https://travis-ci.org/adamhsparks/getCRUCLdata.svg?branch=master)](https://travis-ci.org/)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/adamhsparks/getCRUCLdata?branch=master&svg=true)](https://ci.appveyor.com/project/adamhsparks/getCRUCLdata)
-[![Coverage Status](https://img.shields.io/codecov/c/github/adamhsparks/getCRUCLdata/master.svg)](https://codecov.io/github/adamhsparks/getCRUCLdata?branch=master)
+[![Travis-CI Build Status](https://travis-ci.org/ropensci/getCRUCLdata.svg?branch=master)](https://travis-ci.org/)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/getCRUCLdata?branch=master&svg=true)](https://ci.appveyor.com/project/ropensci/getCRUCLdata)
+[![Coverage Status](https://img.shields.io/codecov/c/github/ropensci/getCRUCLdata/master.svg)](https://codecov.io/github/ropensci/getCRUCLdata?branch=master)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/getCRUCLdata)](https://cran.r-project.org/package=getCRUCLdata)
 [![rstudio mirror downloads](http://cranlogs.r-pkg.org/badges/getCRUCLdata?color=blue)](https://github.com/metacran/cranlogs.app)
 [![rstudio mirror downloads](http://cranlogs.r-pkg.org/badges/grand-total/getCRUCLdata?color=blue)](https://github.com/metacran/cranlogs.app)
@@ -67,7 +67,7 @@ if (!require("devtools")) {
   install.packages("devtools")
 }
 
-devtools::install_github("adamhsparks/getCRUCLdata", build_vignettes = TRUE)
+devtools::install_github("ropensci/getCRUCLdata", build_vignettes = TRUE)
 
 ```
 
@@ -246,7 +246,7 @@ t <- create_CRU_df(tmp = TRUE, dsn = "~/Downloads")
 ------
 # Documentation
 
-For complete documentation see the package website: https://adamhsparks.github.io/getCRUCLdata/
+For complete documentation see the package website: https://ropensci.github.io/getCRUCLdata/
 
 # Meta
 
@@ -272,7 +272,7 @@ Climate Research, 2000, Vol 21, pg 1-25
 
 ## Other
 
-* Please [report any issues or bugs](https://github.com/adamhsparks/getCRUCLdata/issues).
+* Please [report any issues or bugs](https://github.com/ropensci/getCRUCLdata/issues).
 * License: MIT
 * Get citation information for _getCRUCLdata_ in R typing `citation(package = "getCRUCLdata")`
 * Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
