### 2017-04-04 7f2d085
Update devel version of R in cran-comments
