### 2017-04-04 874a0c9
Update Zenodo badge for new 0.1.5 release
